+++ projects/u7chat/src/features/llm-models/model-select.tsx
@@ -1,9 +1,8 @@
 'use client'
 
-import { changeModel } from './litellm-actions'
+import { changeModel } from './llm-model-actions'
 
 export function ModelSelect(props: { models: string[]; defaultModel?: string }) {
-  console.log('Default model:', props.defaultModel)
   return (
     <select
       className='mb-4 w-full rounded border border-gray-300 p-2'
